Separate redux store into it's own file
